Add null check for removeUidFromLocalNetBlockMap Fixes system crash during uninstalling an app: E AndroidRuntime: *** FATAL EXCEPTION IN SYSTEM PROCESS: ConnectivityServiceThread E AndroidRuntime: java.lang.RuntimeException: Error receiving broadcast Intent { act=android.intent.action.PACKAGE_REMOVED dat=package: flg=0x4000010 (has extras) } in android.net.connectivity.com.android.server.connectivity.PermissionMonitor$1@b51a5b5 E AndroidRuntime: 	at android.app.LoadedApk$ReceiverDispatcher$Args.lambda$getRunnable$0(LoadedApk.java:1847) E AndroidRuntime: 	at android.app.LoadedApk$ReceiverDispatcher$Args.$r8$lambda$mcNAAl1SQ4MyJPyDg8TJ2x2h0Rk(Unknown Source:0) E AndroidRuntime: 	at android.app.LoadedApk$ReceiverDispatcher$Args$$ExternalSyntheticLambda0.run(D8$$SyntheticClass:0) E AndroidRuntime: 	at android.os.Handler.handleCallback(Handler.java:995) E AndroidRuntime: 	at android.os.Handler.dispatchMessage(Handler.java:103) E AndroidRuntime: 	at android.os.Looper.loopOnce(Looper.java:248) E AndroidRuntime: 	at android.os.Looper.loop(Looper.java:338) E AndroidRuntime: 	at android.os.HandlerThread.run(HandlerThread.java:85) E AndroidRuntime: Caused by: java.lang.NullPointerException: Attempt to invoke interface method 'boolean android.net.connectivity.com.android.net.module.util.IBpfMap.deleteEntry(android.net.connectivity.com.android.net.module.util.Struct)' on a null object reference E AndroidRuntime: 	at android.net.connectivity.com.android.server.BpfNetMaps.removeUidFromLocalNetBlockMap(BpfNetMaps.java:1074) E AndroidRuntime: 	at android.net.connectivity.com.android.server.connectivity.PermissionMonitor.onPackageRemoved(PermissionMonitor.java:918) E AndroidRuntime: 	at android.net.connectivity.com.android.server.connectivity.PermissionMonitor$1.onReceive(PermissionMonitor.java:183) E AndroidRuntime: 	at android.app.LoadedApk$ReceiverDispatcher$Args.lambda$getRunnable$0(LoadedApk.java:1839) E AndroidRuntime: 	... 7 more

+++ service/src/com/android/server/BpfNetMaps.java
@@ -1070,6 +1070,9 @@ public class BpfNetMaps {
     @RequiresApi(Build.VERSION_CODES.CUR_DEVELOPMENT)
     public void removeUidFromLocalNetBlockMap(final int uid) {
         throwIfPre25Q2("removeUidFromLocalNetBlockMap is not available on pre-B devices");
+
+        if (sLocalNetBlockedUidMap == null) return;
+
         try {
             sLocalNetBlockedUidMap.deleteEntry(new U32(uid));
         } catch (ErrnoException e) {
